Limiting file parsing instances
